refactor(sections): extract flag-row mining into flag_rows.rs
Pure motion: where an option table starts, how one physical line splits
into entries (packed rows, BNF alternations, bracket groups), and where a
nested entry table interrupts a row's description.
